requirements.txt
fastapi==0.104.1
pytest==7.4.3
httpx==0.25.2
uvicorn==0.24.0.post1
pandas==2.1.3
firebase-admin==6.3.0
Pyrebase4==4.7.1
